+++ SCI_Lib/Scripts/Elements/Code.cs
@@ -473,7 +473,7 @@ namespace SCI_Translator.Scripts.Elements
 
         public override void Write(ByteBuilder bb)
         {
-            Address = bb.Position;
+            Address = (ushort)bb.Position;
             bb.AddByte(Type);
             foreach (object arg in Arguments)
             {
